Splitted sankey series into several files.
